+++ src/upload-lighthouse/index.ts
@@ -9,232 +9,15 @@
  * and finally upload this metadata on each macro-run to the appropriate Atlas collection.
  */
 import fs from 'fs';
-import * as github from '@actions/github';
 import { promisify } from 'util';
 import { MongoClient } from 'mongodb';
-import {
-  S3Client,
-  PutObjectCommand,
-  ObjectCannedACL,
-} from '@aws-sdk/client-s3';
 
-const readFileAsync = promisify(fs.readFile);
+import { Manifest } from './types';
+import { DB_NAME, MAIN_COLL_NAME, PR_COLL_NAME } from './constants';
+import { uploadHtmlToS3 } from './uploadToS3';
+import { createRunDocument, sortAndAverageRuns } from './helpers';
 
-/* Summary of important Lighthouse scores of a run already "summarized" by lighthouse library */
-interface Summary {
-  seo: number;
-  performance: number;
-  'best-practices': number;
-  pwa: number;
-  accessibility: number;
-}
-const summaryProperties: (keyof Summary)[] = [
-  'seo',
-  'performance',
-  'best-practices',
-  'pwa',
-  'accessibility',
-];
-
-/* Additional scores to average for DOP purposes */
-interface ExtendedSummary extends Summary {
-  'largest-contentful-paint': number;
-  'first-contentful-paint': number;
-  'total-blocking-time': number;
-  'speed-index': number;
-  'cumulative-layout-shift': number;
-  interactive: number;
-}
-const extendedSummaryProperties: (keyof ExtendedSummary)[] = [
-  'largest-contentful-paint',
-  'first-contentful-paint',
-  'total-blocking-time',
-  'speed-index',
-  'cumulative-layout-shift',
-  'interactive',
-];
-
-/* Manifest structure outputted for each Lighthouse run */
-interface Manifest {
-  url: string;
-  isRepresentativeRun: boolean;
-  htmlPath: string;
-  jsonPath: string;
-  summary: Summary;
-}
-
-/*
- * General type to help define a very large JSON output
- * Documentation of JSON output: https://github.com/GoogleChrome/lighthouse/blob/main/docs/understanding-results.md
- */
-interface JsonRun {
-  [k: string]: unknown;
-  audits: {
-    [k in keyof ExtendedSummary]: {
-      [k: string]: unknown;
-      score: number;
-    };
-  };
-}
-
-interface RunDocument {
-  summary: ExtendedSummary;
-  url: string;
-  type: 'desktop' | 'mobile';
-  commitHash: string;
-  commitMessage: string;
-  commitTimestamp: Date;
-  author: string;
-  project: string;
-  branch: string;
-}
-
-const DB_NAME = `lighthouse`;
-/* Used on PR creation and update (synchronize) */
-const PR_COLL_NAME = `pr_reports`;
-/* Used on merge to main in Snooty to keep running scores of production */
-const MAIN_COLL_NAME = `main_reports`;
-
-/* Helpers */
-const getEmptySummary = (): ExtendedSummary => ({
-  seo: 0,
-  performance: 0,
-  'best-practices': 0,
-  pwa: 0,
-  accessibility: 0,
-  'largest-contentful-paint': 0,
-  'first-contentful-paint': 0,
-  'speed-index': 0,
-  interactive: 0,
-  'total-blocking-time': 0,
-  'cumulative-layout-shift': 0,
-});
-
-const getAverageSummary = (
-  manifests: Manifest[],
-  jsonRuns: JsonRun[],
-): ExtendedSummary => {
-  const summary = getEmptySummary();
-  for (const property of summaryProperties) {
-    summary[property] =
-      manifests.reduce((acc, cur) => acc + cur.summary[property], 0) /
-      manifests.length;
-  }
-  for (const property of extendedSummaryProperties) {
-    summary[property] =
-      jsonRuns.reduce((acc, cur) => acc + cur.audits[property].score, 0) /
-      jsonRuns.length;
-  }
-  return summary;
-};
-
-/* Reads and returns files of runs in arrays */
-const getRuns = async (
-  manifests: Manifest[],
-): Promise<{ jsonRuns: JsonRun[]; htmlRuns: string[] }> => {
-  const jsonRuns = [];
-  const htmlRuns = [];
-
-  for (const manifest of manifests) {
-    jsonRuns.push(
-      JSON.parse((await readFileAsync(manifest.jsonPath)).toString()),
-    );
-
-    htmlRuns.push((await readFileAsync(manifest.htmlPath)).toString());
-  }
-
-  await Promise.all(jsonRuns);
-  await Promise.all(htmlRuns);
-  return { jsonRuns, htmlRuns };
-};
-
-interface SortedRuns {
-  htmlRuns: string[];
-  summary: ExtendedSummary;
-  url: string;
-}
-
-const sortAndAverageRuns = async (
-  manifests: Manifest[],
-): Promise<SortedRuns[]> => {
-  const runs: {
-    htmlRuns: string[];
-    summary: ExtendedSummary;
-    url: string;
-  }[] = [];
-  const uniqueUrls = new Set(manifests.map(manifest => manifest.url));
-
-  for (const url of uniqueUrls) {
-    const manifestsForUrl = manifests.filter(manifest => manifest.url === url);
-    const { jsonRuns, htmlRuns } = await getRuns(manifestsForUrl);
-    const summary = getAverageSummary(manifestsForUrl, jsonRuns);
-    runs.push({ htmlRuns, summary, url });
-  }
-
-  return runs;
-};
-
-const createRunDocument = (
-  { url, summary }: SortedRuns,
-  type: 'mobile' | 'desktop',
-): RunDocument => {
-  const commitHash = github.context.sha;
-  const author = github.context.actor;
-  const commitMessage = process.env.COMMIT_MESSAGE || '';
-  const commitTimestamp = process.env.COMMIT_TIMESTAMP
-    ? new Date(process.env.COMMIT_TIMESTAMP)
-    : new Date();
-  const project = process.env.PROJECT_TO_BUILD || '';
-  const branch = process.env.BRANCH_NAME || '';
-
-  return {
-    commitHash,
-    commitMessage,
-    commitTimestamp,
-    author,
-    project,
-    branch,
-    url,
-    summary,
-    type,
-  };
-};
-
-// Html reports will be uploaded to S3. The 
-async function uploadHtmlToS3(
-  { htmlRuns, url, }: SortedRuns,
-  type: 'mobile' | 'desktop'
-) {
-  const AWS_BUCKET = 'docs-lighthouse';
-  const commitHash = github.context.sha;
-  const branch = process.env.BRANCH_NAME || '';
-  const client = new S3Client();
-
-  const reportType = branch === 'main' ? 'main_reports': 'pr_reports';
-
-  let cleanedUrl = url.replace('http://localhost:9000/', '');
-  if (cleanedUrl.endsWith('?desktop')) cleanedUrl = cleanedUrl.slice(0, -8);
-  cleanedUrl = cleanedUrl.split(/\/\/|\//).join('-');
-
-  const destinationDir = `${reportType}/${commitHash}/${cleanedUrl}/${type}`;
-
-  const uploads = htmlRuns.map(async (htmlReport, i) => {
-    const key = `${destinationDir}/${i + 1}.html`;
-    console.log('Uploading to S3 at ', key);
-
-    const input = {
-      Body: Buffer.from(htmlReport),
-      Key: key,
-      Bucket: AWS_BUCKET,
-      ContentType: 'text/html',
-      CacheControl: 'no-cache',
-      ACL: ObjectCannedACL.public_read,
-    };
-    const command = new PutObjectCommand(input);
-    return client.send(command);
-  });
-  return Promise.all(uploads);
-};
+export const readFileAsync = promisify(fs.readFile);
 
 async function main(): Promise<void> {
   const branch = process.env.BRANCH_NAME || '';
